extract hilbert, tests, attribution
