+++ planetiler-core/src/main/java/com/onthegomap/planetiler/geo/TileCoord.java
@@ -21,8 +21,34 @@ import org.locationtech.jts.geom.CoordinateXY;
  */
 @Immutable
 public record TileCoord(int encoded, int x, int y, int z) implements Comparable<TileCoord> {
+
+  private static final int MAX_ZOOM = 15;
+
+  private static final int[] ZOOM_START_INDEX = new int[MAX_ZOOM + 1];
+
+  static {
+    int idx = 0;
+    for (int z = 0; z <= MAX_ZOOM; z++) {
+      ZOOM_START_INDEX[z] = idx;
+      idx += (1 << z) * (1 << z);
+    }
+  }
+
+  private static int startIndexForZoom(int z) {
+    return ZOOM_START_INDEX[z];
+  }
+
+  private static int zoomForIndex(int idx) {
+    for (int z = MAX_ZOOM; z >= 0; z--) {
+      if (ZOOM_START_INDEX[z] <= idx) {
+        return z;
+      }
+    }
+    throw new IllegalArgumentException("Bad index: " + idx);
+  }
+
   public TileCoord {
-    assert z <= 15;
+    assert z <= MAX_ZOOM;
   }
 
   public static TileCoord ofXYZ(int x, int y, int z) {
@@ -30,19 +56,9 @@ public record TileCoord(int encoded, int x, int y, int z) implements Comparable<
   }
 
   public static TileCoord decode(int encoded) {
-    int acc = 0;
-    int tmpZ = 0;
-    while (true) {
-      int numTiles = (1 << tmpZ) * (1 << tmpZ);
-      if (acc + numTiles > encoded) {
-        int position = encoded - acc;
-        // long xy = hilbertPositionToXY(tmpZ, position);
-        long xy = tmsPositionToXY(tmpZ, position);
-        return new TileCoord(encoded, (int) (xy >>> 32 & 0xFFFFFFFFL), (int) (xy & 0xFFFFFFFFL), tmpZ);
-      }
-      acc += numTiles;
-      tmpZ++;
-    }
+    int z = zoomForIndex(encoded);
+    long xy = tmsPositionToXY(z, encoded - startIndexForZoom(z));
+    return new TileCoord(encoded, (int) (xy >>> 32 & 0xFFFFFFFFL), (int) (xy & 0xFFFFFFFFL), z);
   }
 
   /** Returns the tile containing a latitude/longitude coordinate at a given zoom level. */
@@ -54,12 +70,7 @@ public record TileCoord(int encoded, int x, int y, int z) implements Comparable<
   }
 
   public static int encode(int x, int y, int z) {
-    int acc = 0;
-    for (int tmpZ = 0; tmpZ < z; tmpZ++) {
-      acc += (1 << tmpZ) * (1 << tmpZ);
-    }
-    // return acc + hilbertXYToPosition(z, x, y);
-    return acc + tmsXYToPosition(z, x, y);
+    return startIndexForZoom(z) + tmsXYToPosition(z, x, y);
   }
 
   @Override
@@ -86,17 +97,11 @@ public record TileCoord(int encoded, int x, int y, int z) implements Comparable<
     return "{x=" + x + " y=" + y + " z=" + z + '}';
   }
 
-  public double progressOnLevel() {
-    int acc = 0;
-    int tmpZ = 0;
-    while (true) {
-      int numTiles = (1 << tmpZ) * (1 << tmpZ);
-      if (acc + numTiles > encoded) {
-        return (encoded - acc) / (double) numTiles;
-      }
-      acc += numTiles;
-      tmpZ++;
-    }
+  public double progressOnLevel(TileExtents extents) {
+    // approximate percent complete within a bounding box by computing what % of the way through the columns we are
+    // (for hilbert ordering, we probably won't be able to reflect the bounding box)
+    var zoomBounds = extents.getForZoom(z);
+    return 1d * (x - zoomBounds.minX()) / (zoomBounds.maxX() - zoomBounds.minX());
   }
 
   @Override
@@ -140,108 +145,5 @@ public record TileCoord(int encoded, int x, int y, int z) implements Comparable<
   public static int tmsXYToPosition(int z, int x, int y) {
     int dim = 1 << z;
     return x * dim + (dim - 1 - y);
-  }
-
-  // hilbert implementation (not currently used)
-  // Fast Hilbert curve algorithm by http://threadlocalmutex.com/
-  // Ported from C++ https://github.com/rawrunprotected/hilbert_curves (public domain)
-  private static int deinterleave(int tx) {
-    tx = tx & 0x55555555;
-    tx = (tx | (tx >>> 1)) & 0x33333333;
-    tx = (tx | (tx >>> 2)) & 0x0F0F0F0F;
-    tx = (tx | (tx >>> 4)) & 0x00FF00FF;
-    tx = (tx | (tx >>> 8)) & 0x0000FFFF;
-    return tx;
-  }
-
-  private static int interleave(int tx) {
-    tx = (tx | (tx << 8)) & 0x00FF00FF;
-    tx = (tx | (tx << 4)) & 0x0F0F0F0F;
-    tx = (tx | (tx << 2)) & 0x33333333;
-    tx = (tx | (tx << 1)) & 0x55555555;
-    return tx;
-  }
-
-  private static int prefixScan(int tx) {
-    tx = (tx >>> 8) ^ tx;
-    tx = (tx >>> 4) ^ tx;
-    tx = (tx >>> 2) ^ tx;
-    tx = (tx >>> 1) ^ tx;
-    return tx;
-  }
-
-  private static long hilbertPositionToXY(int z, int pos) {
-    pos = pos << (32 - 2 * z);
-
-    int i0 = deinterleave(pos);
-    int i1 = deinterleave(pos >>> 1);
-
-    int t0 = (i0 | i1) ^ 0xFFFF;
-    int t1 = i0 & i1;
-
-    int prefixT0 = prefixScan(t0);
-    int prefixT1 = prefixScan(t1);
-
-    int a = (((i0 ^ 0xFFFF) & prefixT1) | (i0 & prefixT0));
-
-    int resultX = (a ^ i1) >>> (16 - z);
-    int resultY = (a ^ i0 ^ i1) >>> (16 - z);
-    return ((long) resultX << 32) | resultY;
-  }
-
-  private static int hilbertXYToIndex(int z, int x, int y) {
-    x = x << (16 - z);
-    y = y << (16 - z);
-
-    int hA, hB, hC, hD;
-
-    int a1 = x ^ y;
-    int b1 = 0xFFFF ^ a1;
-    int c1 = 0xFFFF ^ (x | y);
-    int d1 = x & (y ^ 0xFFFF);
-
-    hA = a1 | (b1 >>> 1);
-    hB = (a1 >>> 1) ^ a1;
-
-    hC = ((c1 >>> 1) ^ (b1 & (d1 >>> 1))) ^ c1;
-    hD = ((a1 & (c1 >>> 1)) ^ (d1 >>> 1)) ^ d1;
-
-    int a2 = hA;
-    int b2 = hB;
-    int c2 = hC;
-    int d2 = hD;
-
-    hA = ((a2 & (a2 >>> 2)) ^ (b2 & (b2 >>> 2)));
-    hB = ((a2 & (b2 >>> 2)) ^ (b2 & ((a2 ^ b2) >>> 2)));
-
-    hC ^= ((a2 & (c2 >>> 2)) ^ (b2 & (d2 >>> 2)));
-    hD ^= ((b2 & (c2 >>> 2)) ^ ((a2 ^ b2) & (d2 >>> 2)));
-
-    int a3 = hA;
-    int b3 = hB;
-    int c3 = hC;
-    int d3 = hD;
-
-    hA = ((a3 & (a3 >>> 4)) ^ (b3 & (b3 >>> 4)));
-    hB = ((a3 & (b3 >>> 4)) ^ (b3 & ((a3 ^ b3) >>> 4)));
-
-    hC ^= ((a3 & (c3 >>> 4)) ^ (b3 & (d3 >>> 4)));
-    hD ^= ((b3 & (c3 >>> 4)) ^ ((a3 ^ b3) & (d3 >>> 4)));
-
-    int a4 = hA;
-    int b4 = hB;
-    int c4 = hC;
-    int d4 = hD;
-
-    hC ^= ((a4 & (c4 >>> 8)) ^ (b4 & (d4 >>> 8)));
-    hD ^= ((b4 & (c4 >>> 8)) ^ ((a4 ^ b4) & (d4 >>> 8)));
-
-    int a = hC ^ (hC >>> 1);
-    int b = hD ^ (hD >>> 1);
-
-    int i0 = x ^ y;
-    int i1 = b | (0xFFFF ^ (i0 | a));
-
-    return ((interleave(i1) << 1) | interleave(i0)) >>> (32 - 2 * z);
   }
 }
