+++ planetiler-core/src/main/java/com/onthegomap/planetiler/mbtiles/MbtilesWriter.java
@@ -205,7 +205,7 @@ public class MbtilesWriter {
     } else {
       blurb = "%d/%d/%d (z%d %s%%) %s".formatted(
         lastTile.z(), lastTile.x(), lastTile.y(),
-        lastTile.z(), 100 * lastTile.progressOnLevel(),
+        lastTile.z(), 100 * lastTile.progressOnLevel(config.bounds().tileExtents()),
         lastTile.getDebugUrl()
       );
     }
